Fix backup command by using snapshot upload
Backup does not work because files are too big to not use multipart
uploads. This fixes a key command by factoring out the upload code to be
used by backup.
