+++ ironfish-cli/src/commands/service/snapshot.ts
@@ -1,26 +1,17 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import {
-  AbortMultipartUploadCommand,
-  CompleteMultipartUploadCommand,
-  CreateMultipartUploadCommand,
-  S3Client,
-  UploadPartCommand,
-} from '@aws-sdk/client-s3'
-import { Assert, FileUtils, NodeUtils } from '@ironfish/sdk'
+import { S3Client } from '@aws-sdk/client-s3'
+import { FileUtils, NodeUtils } from '@ironfish/sdk'
 import { CliUx, Flags } from '@oclif/core'
 import crypto from 'crypto'
 import fsAsync from 'fs/promises'
 import path from 'path'
-import tar from 'tar'
 import { IronfishCommand } from '../../command'
 import { LocalFlags } from '../../flags'
-import { DEFAULT_SNAPSHOT_BUCKET, SnapshotManifest } from '../../utils'
+import { DEFAULT_SNAPSHOT_BUCKET, S3Utils, SnapshotManifest } from '../../utils'
+import { TarUtils } from '../../utils/tar'
 
-// AWS requires that upload parts be at least 5MB
-const MINIMUM_MULTIPART_FILE_SIZE = 5 * 1024 * 1024
-const MAX_MULTIPART_NUM = 10000
 const SNAPSHOT_FILE_NAME = `ironfish_snapshot.tar.gz`
 
 export default class CreateSnapshot extends IronfishCommand {
@@ -99,7 +90,7 @@ export default class CreateSnapshot extends IronfishCommand {
 
     this.log(`Zipping\n    SRC ${chainDatabasePath}\n    DST ${snapshotPath}\n`)
     CliUx.ux.action.start(`Zipping ${chainDatabasePath}`)
-    await this.zipDir(chainDatabasePath + '/', snapshotPath)
+    await TarUtils.zipDir(chainDatabasePath + '/', snapshotPath)
     const stat = await fsAsync.stat(snapshotPath)
     const fileSize = stat.size
     CliUx.ux.action.stop(`done (${FileUtils.formatFileSize(fileSize)})`)
@@ -118,15 +109,23 @@ export default class CreateSnapshot extends IronfishCommand {
     if (flags.upload) {
       const snapshotBaseName = path.basename(SNAPSHOT_FILE_NAME, '.tar.gz')
       const snapshotKeyName = `${snapshotBaseName}_${timestamp}.tar.gz`
+
+      const s3 = new S3Client({
+        credentials: {
+          accessKeyId,
+          secretAccessKey,
+        },
+        region,
+      })
+
       CliUx.ux.action.start(`Uploading to ${bucket}`)
-      await this.uploadToBucket(
+      await S3Utils.uploadToBucket(
+        s3,
         snapshotPath,
         'application/x-compressed-tar',
         bucket,
         snapshotKeyName,
-        accessKeyId,
-        secretAccessKey,
-        region,
+        this.logger.withTag('s3'),
       )
       CliUx.ux.action.stop(`done`)
 
@@ -144,14 +143,13 @@ export default class CreateSnapshot extends IronfishCommand {
         .writeFile(manifestPath, JSON.stringify(manifest, undefined, '  '))
         .then(async () => {
           CliUx.ux.action.start(`Uploading latest snapshot information to ${bucket}`)
-          await this.uploadToBucket(
+          await S3Utils.uploadToBucket(
+            s3,
             manifestPath,
             'application/json',
             bucket,
             manifestPath,
-            accessKeyId,
-            secretAccessKey,
-            region,
+            this.logger.withTag('s3'),
           )
           CliUx.ux.action.stop(`done`)
         })
@@ -159,182 +157,5 @@ export default class CreateSnapshot extends IronfishCommand {
       this.log('Snapshot upload complete. Uploaded the following manifest:')
       this.log(JSON.stringify(manifest, undefined, '  '))
     }
-  }
-
-  async zipDir(source: string, dest: string, excludes: string[] = []): Promise<void> {
-    const sourceDir = path.dirname(source)
-    const sourceFile = path.basename(source)
-    const excludeSet = new Set(excludes)
-
-    await tar.create(
-      {
-        gzip: true,
-        file: dest,
-        C: sourceDir,
-        filter: function (path) {
-          if (excludeSet.has(path)) {
-            return false
-          } else {
-            return true
-          }
-        },
-      },
-      [sourceFile],
-    )
-  }
-
-  async uploadToBucket(
-    filePath: string,
-    contentType: string,
-    bucket: string,
-    keyName: string,
-    accessKeyId: string,
-    secretAccessKey: string,
-    region: string,
-  ): Promise<void> {
-    const fileHandle = await fsAsync.open(filePath, 'r')
-
-    const stat = await fsAsync.stat(filePath)
-    const fileSize = stat.size
-    let numParts = MAX_MULTIPART_NUM
-
-    while (fileSize / numParts < MINIMUM_MULTIPART_FILE_SIZE) {
-      numParts /= 2
-    }
-
-    const uploadChunkSize = fileSize / numParts
-
-    const contentStream = fileHandle.createReadStream()
-
-    const s3 = new S3Client({
-      credentials: {
-        accessKeyId,
-        secretAccessKey,
-      },
-      region,
-    })
-
-    const params = {
-      Bucket: bucket,
-      Key: keyName,
-      ContentType: contentType,
-    }
-
-    const uploadId = await s3
-      .send(new CreateMultipartUploadCommand(params))
-      .then((result) => result.UploadId)
-      .catch((err: Error) => {
-        this.logger.error(`Could not create multipart upload to S3: ${err.message}`)
-        throw new Error(err.message)
-      })
-
-    Assert.isNotUndefined(uploadId)
-
-    const uploadPartsPromise = new Promise<{
-      Parts: { ETag: string | undefined; PartNumber: number }[]
-    }>((resolve, reject) => {
-      const partMap: { Parts: { ETag: string | undefined; PartNumber: number }[] } = {
-        Parts: [],
-      }
-
-      let partNum = 1
-      let acc: Buffer | null = null
-
-      contentStream.on('data', (chunk: Buffer) => {
-        if (!acc) {
-          acc = chunk
-        } else {
-          acc = Buffer.concat([acc, chunk])
-        }
-
-        if (acc.length > uploadChunkSize) {
-          contentStream.pause()
-
-          const params = {
-            Bucket: bucket,
-            Key: keyName,
-            PartNumber: partNum,
-            UploadId: uploadId,
-            ContentType: contentType,
-            Body: acc,
-          }
-
-          s3.send(new UploadPartCommand(params))
-            .then((result) => {
-              partMap.Parts.push({ ETag: result.ETag, PartNumber: params.PartNumber })
-              partNum += 1
-              acc = null
-              contentStream.resume()
-            })
-            .catch((err: Error) => {
-              this.logger.error(`Could not upload part to S3 bucket: ${err.message}`)
-              reject(err)
-            })
-        }
-      })
-
-      contentStream.on('close', () => {
-        if (acc) {
-          const params = {
-            Bucket: bucket,
-            Key: keyName,
-            PartNumber: partNum,
-            UploadId: uploadId,
-            ContentType: contentType,
-            Body: acc,
-          }
-
-          s3.send(new UploadPartCommand(params))
-            .then((result) => {
-              partMap.Parts.push({ ETag: result.ETag, PartNumber: params.PartNumber })
-              acc = null
-              resolve(partMap)
-            })
-            .catch((err: Error) => {
-              this.logger.error(`Could not upload last part to S3 bucket: ${err.message}`)
-              reject(err)
-            })
-        }
-      })
-
-      contentStream.on('error', (err) => {
-        this.logger.error(`Could not read file: ${err.message}; aborting upload to S3...`)
-        const params = {
-          Bucket: bucket,
-          Key: keyName,
-          UploadId: uploadId,
-        }
-
-        s3.send(new AbortMultipartUploadCommand(params))
-          .then()
-          .catch((awsErr: Error) => {
-            this.logger.error(`Could not abort S3 upload: ${awsErr.message}`)
-          })
-
-        reject(err)
-      })
-    })
-
-    const partMap = await uploadPartsPromise
-
-    this.logger.debug(
-      `All parts of snapshot have been uploaded. Finalizing multipart upload. Parts: ${partMap.Parts.length}`,
-    )
-
-    const completionParams = {
-      Bucket: bucket,
-      Key: keyName,
-      UploadId: uploadId,
-      MultipartUpload: partMap,
-    }
-
-    await s3
-      .send(new CompleteMultipartUploadCommand(completionParams))
-      .then(() => {
-        this.logger.info(`Multipart upload complete.`)
-      })
-      .catch((err: Error) => {
-        throw new Error(`Could not complete multipart S3 upload: ${err.message}`)
-      })
   }
 }
