+++ ironfish-cli/src/utils/index.ts
@@ -3,6 +3,8 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 export * from './editor'
 export * from './rpc'
+export * from './s3'
 export * from './snapshot'
+export * from './tar'
 export * from './terminal'
 export * from './types'
